requirements-dev.txt
asgiref==3.2.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
django-imagekit==4.0.2